MAJOR PATCH: Fixed CMap functionality and added thresholds
I fixed the major bug of the CMap functionality which included a mistake
in setting the advanced for-loop. Also added a threshold variable via
method overloading to allow for a "reasonable fit" solution to GREATLY
speedup processing at the cost of reduction in quality. Included my
local binaries and a new source file image. Lena.jpg is raw and Lean.png
is the one that's been processed to grayscale, i.e. use the Lena.png for
encoding/decoding purposes