Multithreading branch comments added + Threshold Debugging
Added comments to the various parts of the EncodeThread class.

Also added comments to the Multithreading-specific methods and
alterations in IFSEncoder

Debugging goes poorly due to last-minute alterations. I have set up the
main method to encode the image with threads and without threads. It
seems to be faster without threads, possibly due to the current
threshold setting and image size. Or maybe it's my processor on
battery-saver mode.

Regardless, I've left debugging features in because of the test-build
nature of this build.